requirements.txt
pygame==2.5.2
python-chess==1.999
numpy>=1.26.0
scikit-learn>=1.4.0
joblib>=1.3.2
scipy>=1.12.0 
